requirements.txt
absl-py==1.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
comm==0.1.4
coremltools==3.3
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.22
entrypoints==0.4
Flask==2.0.3
Flask-MySQL==1.5.2
flask-redis==0.4.0
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.10.0
idna==3.4
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.8.1
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
llvmlite==0.33.0
Markdown==3.3.7
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.8
notebook==6.4.10
numba==0.50.1
numpy==1.18.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prettytable==0.7.2
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
PyMySQL==1.0.2
pyparsing==3.1.1
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3.post1
pyzmq==25.1.1
qtconsole==5.2.2
QtPy==2.0.1
redis==4.3.6
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.1
rsa==4.9
scipy==1.5.4
Send2Trash==1.8.2
six==1.16.0
sqlparse==0.4.4
tensorboard==2.0.2
tensorflow==2.0.4
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
tornado==6.1
traitlets==4.3.3
turicreate==6.4.1
typing_extensions==4.1.1
urllib3==1.26.18
wcwidth==0.2.8
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.6
wrapt==1.15.0
zipp==3.6.0
